# planning-service/openapi.json
{
  "components": {
    "schemas": {
...
      "Classification": {
        "additionalProperties": true,
        "description": "What the chart draws at one point; None where it has no coverage.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "category"
        ],
        "title": "Classification",
        "type": "object"
      },
...
      "DisplacedPick": {
        "description": "A pick on another point that a save took the place of: the store\nkeeps one pick per place (vfr.routecsv.SAME_PLACE_NM), whatever each\npoint is.",
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          }
        },
        "required": [
          "lat",
          "lon",
          "category"
        ],
        "title": "DisplacedPick",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Pick": {
        "description": "A pilot's own mark on the chart. `rated` is derived from the\nrating, so a rated point can never describe itself as unrated.",
        "properties": {
          "along_track_nm": {
            "title": "Along Track Nm",
            "type": "number"
          },
          "area_m2": {
            "default": 0.0,
            "title": "Area M2",
            "type": "number"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Created At"
          },
          "cross_track_nm": {
            "title": "Cross Track Nm",
            "type": "number"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Note"
          },
          "rated": {
            "readOnly": true,
            "title": "Rated",
            "type": "boolean"
          },
          "rating": {
            "anyOf": [
              {
                "enum": [
                  0,
                  1,
                  2,
                  3,
                  4,
                  5
                ],
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Rating"
          },
          "role": {
            "anyOf": [
              {
                "enum": [
                  "dr",
                  "visual"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Role"
          },
          "route": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Route"
          },
          "source": {
            "enum": [
              "detected",
              "added"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source",
          "category",
          "lat",
          "lon",
          "along_track_nm",
          "cross_track_nm",
          "rated"
        ],
        "title": "Pick",
        "type": "object"
      },
      "PickDeleted": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "summary": {
            "$ref": "#/components/schemas/PickSummary"
          }
        },
        "required": [
          "ok",
          "summary"
        ],
        "title": "PickDeleted",
        "type": "object"
      },
      "PickRequest": {
        "properties": {
          "area_m2": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area M2"
          },
          "category": {
            "default": "water",
            "title": "Category",
            "type": "string"
          },
          "departure_ident": {
            "title": "Departure Ident",
            "type": "string"
          },
          "destination_ident": {
            "title": "Destination Ident",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "source": {
            "default": "detected",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "departure_ident",
          "destination_ident",
          "lat",
          "lon"
        ],
        "title": "PickRequest",
        "type": "object"
      },
      "PickSaved": {
        "properties": {
          "displaced": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DisplacedPick"
            },
            "title": "Displaced",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "pick": {
            "$ref": "#/components/schemas/Pick"
          },
          "summary": {
            "$ref": "#/components/schemas/PickSummary"
          }
        },
        "required": [
          "ok",
          "pick",
          "summary"
        ],
        "title": "PickSaved",
        "type": "object"
      },
      "PickSummary": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "integer"
          },
          "added": {
            "title": "Added",
            "type": "integer"
          },
          "added_categories": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Added Categories",
            "type": "array"
          },
          "by_rating": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Rating",
            "type": "object"
          },
          "by_role": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Role",
            "type": "object"
          },
          "rejected": {
            "title": "Rejected",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "accepted",
          "rejected",
          "added",
          "by_rating",
          "by_role"
        ],
        "title": "PickSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Wingtip Maps planner",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/chart-tile/{kind}/{z}/{x}/{y}.png": {
      "get": {
        "description": "Any kind of chart by its key (`chart_layers` on the course lists\nthem): `sec`, `tac`, `ifr_low`, `ifr_high`, `ifr_area`. What the\nmap's layers draw from.\n\nThe FAA's own GeoTIFF of each sheet, rendered into a {z}/{x}/{y} tile\npyramid (vfr.charts -- downloaded once per 56-day chart cycle) and\ncached on disk; the same tiles the detector reads. A tile pyramid is\nwhat makes each chart a plain Leaflet tile layer: edge-only fetches\non a pan, the previous zoom's tiles scaled under the zoom animation,\na prefetch ring. 404 where no sheet covers the tile, which the layer\nleaves blank. The sectional and the TAC had routes of their own\nunder older names, which nothing called any more.",
        "operationId": "chart_tile_api_chart_tile__kind___z___x___y__png_get",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "z",
            "required": true,
            "schema": {
              "title": "Z",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "x",
            "required": true,
            "schema": {
              "title": "X",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "y",
            "required": true,
            "schema": {
              "title": "Y",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chart Tile"
      }
    },
...
    "/api/classify": {
      "get": {
        "description": "What the chart draws at a point, so a hand-marked checkpoint is\ncategorised from the pixels rather than from whatever the dropdown\nhappened to be left on.",
        "operationId": "classify_api_classify_get",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "lon",
            "required": true,
            "schema": {
              "title": "Lon",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Classification"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Classify"
      }
    },
...
    "/api/detect/stream": {
      "get": {
        "description": "Detections as newline-delimited JSON, one line per block; each\nline is one app.schemas.DetectMessage.\n\nBlocks advance along the course, so the map fills from the departure\nend and a pilot can start judging the first thirty miles while the\nrest of the corridor is still being read. Airports come first, in\ntheir own line, because they are a local file lookup and cost\nnothing to produce.\n\nThe landmarks come from the shared per-route job (app.detection);\nonly the pick-matching is per-request, because picks change between\nrequests and the chart does not.\n\nEach carries the chart model's score (app.chart_model) where the\ncorridor has been read whole already, which is every time but a\nroute's first: a detection's features count its neighbours, and a\nblock still being read does not have them all yet.",
        "operationId": "detect_stream_api_detect_stream_get",
        "parameters": [
          {
            "in": "query",
            "name": "dep",
            "required": true,
            "schema": {
              "title": "Dep",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dest",
            "required": true,
            "schema": {
              "title": "Dest",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "half_width_nm",
            "required": false,
            "schema": {
              "default": 4.0,
              "title": "Half Width Nm",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detect Stream"
      }
    },
...
    "/api/picks": {
      "delete": {
        "operationId": "remove_pick_api_picks_delete",
        "parameters": [
          {
            "in": "query",
            "name": "dep",
            "required": true,
            "schema": {
              "title": "Dep",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dest",
            "required": true,
            "schema": {
              "title": "Dest",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "lon",
            "required": true,
            "schema": {
              "title": "Lon",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickDeleted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Pick"
      },
      "post": {
        "operationId": "add_pick_api_picks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PickRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickSaved"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Pick"
      }
    },
...
  }
}